move WebSecurityConfigurerAdapter recipe to security5
