+++ src/main/resources/META-INF/rewrite/spring-boot-27.yml
@@ -37,7 +37,6 @@ recipeList:
       retainVersions:
         - mysql:mysql-connector-java
   # Use recommended replacements for deprecated APIs
-  - org.openrewrite.java.spring.boot2.WebSecurityConfigurerAdapter
   - org.openrewrite.java.ChangeType:
       oldFullyQualifiedTypeName: org.springframework.boot.web.server.LocalServerPort
       newFullyQualifiedTypeName: org.springframework.boot.test.web.server.LocalServerPort
