+++ src/testWithSpringSecurity_5_7/java/org/openrewrite/spring/security5/WebSecurityConfigurerAdapterTest.java
@@ -13,11 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.openrewrite.java.spring.boot2;
+package org.openrewrite.spring.security5;
 
 import org.junit.jupiter.api.Test;
 import org.openrewrite.DocumentExample;
 import org.openrewrite.java.JavaParser;
+import org.openrewrite.java.spring.security5.WebSecurityConfigurerAdapter;
 import org.openrewrite.test.RecipeSpec;
 import org.openrewrite.test.RewriteTest;
